+++ CHANGELOG.md
@@ -22,11 +22,21 @@ Given a version number `MAJOR.MINOR.PATCH`, we increment the:
 
 ## Unreleased
 
+### Added
+
+- Add telemetry to collect anonymous usage metrics.
+  - This helps us to improve user experience by measuring which Terramate features are used most actively.
+    For further details, see [documentation](https://terramate.io/docs/cli/telemetry).
+  - Can be turned off by setting `terramate.config.telemetry.enabled = false` in the project configuration,
+    or by setting `disable_telemetry = true` in the user configuration.
+
 ### Fixed
 
 - Fix the command-line parsing of `run` and `script run` which were not failing from unknown flags.
 - Fix `create --all-terragrunt` creating Terragrunt stacks with cycles.
 - Improve the error reporting of the Outputs Sharing feature.
+- Fix crash in the Terragrunt integration when the project had modules with dependency paths outside the current Terramate project.
+  - A warning will be shown for such configurations.
 
 ## v0.11.3